backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
geopy
g4f
google-genai
